Refactored viewCandidate routes
