+++ api/routes/index.js
@@ -1,50 +1,9 @@
 var models  = require('../models');
 var express = require('express');
-var Sequelize = require('sequelize');
 var router  = express.Router();
 
-var sequelize = new Sequelize('main', 'user', 'password', {
-  dialect: 'sqlite',
-  storage: 'db.development.sqlite'
-});
+var candidateInfoRoutes = require('./candidateInfo');
 
-// Insert a candidate
-router.post('/insert', function(req, res) {
-  // create an instance
-  var candidate = models.candidateInfo.build({
-    name: req.body.name,
-    email: req.body.email,
-    phoneNumber: req.body.phoneNumber,
-    school: req.body.school,
-    major: req.body.major,
-    graduationDate: req.body.graduationDate,
-    needSponsorship: req.body.needSponsorship,
-    internOrFull: req.body.internOrFull,
-    areaOfInterest: req.body.areaOfInterest,
-    preferredLanguages: req.body.preferredLanguages,
-    finalEvaluation: req.body.finalEvaluation
-  })
-  // persist an instance
-  candidate.save().then(() => {
-    res.json(candidate);
-  });
-});
-
-// Get all candidates
-router.get('/getCandidateInfo', function(req, res) {
-  models.candidateInfo.findAll({})
-  .then(function(result) {
-    res.json(result);
-  });
-});
-
-// Get candidate by id
-router.get('/getCandidateInfo/:id', function(req, res) {
-  var id = req.params.id;
-
-  models.candidateInfo.findById(id).then(function(result) {
-    res.json(result);
-  });
-});
+router.use('/candidateInfo', candidateInfoRoutes);
 
 module.exports = router;
